Remove MemoryType trait
In the end, it is WAY simpler to just duplicate the code.  I don't like
the lack of inheritance in Rust, but this is the best way to do it for
now.  The boilerplate and extra steps for trait composition added way
more complexity than it saved in this case.