Add cache miss replacement
Cache can now refresh itself from memory in the case of a miss.  Uses
a random replacement policy, which is fine for now but may be changed in
the future.

God help the read-from-lower implementation, it's really ugly but it
works.  Something to rework in the future once I get a bit better at
Rust.