Simplify memory module
Memory had gotten significantly overcomplicated because I was trying to
optimize and use references before having a basic plan in place.  Clone
is your friend.

Everything now in place and fixed except for the connection between the
levels of memory.  Once that is done, I'll need to write a brief testing
suite to ensure everything is doing what I want it to.