Rework memory write
Write only needs a bool, not an option, it can never return a value.

Also reworked the "random" replacement policy so it will always replace
the same location when it gets the same line.  Not great, but I think it
will be easier for now when trying to do cache replacement.